support multiple --url flags in the CLI
This was necessary to mirror a certain site because the crawler could
not crawl the nav.  So I would have had to do s separate crawl for each
nav section, then a big warc merge or something.  Would have looked like
repeated requests for the same urls in a short period of time!

With this commit, I was able to feed the navigation hints to browsertrix
by using a bunch of --url calls, like:

./browsertrix-crawler.sh \
    ratin-amin_fr \
    'http://ramtin-amin.fr/#tristar' \
    'http://ramtin-amin.fr/#femto_ubi' \
    'http://ramtin-amin.fr/#femto_ubi2' \
    'http://ramtin-amin.fr/#femto_voda' \
    'http://ramtin-amin.fr/#bga_unsolder'